Total Revamp of database
At a stopping point.  The database has been redone so that the rp2
pc_view records can be used more easily and logically.  That caused
some of the maps to be redone.  Some other maps were deemed not useful
and were eliminated.  All the test fixtures, unit tests, and
functional tests were deleted because they were just unused
skeletons.  A few were added back as a start to testing.  This
checking is not functional in any sense but is a checkpoint for the
day.