I got interrupted by ActiveRecord Constraints.
Middle of creating tests, someone asked for a copy of ActiveRecord
Constraints so I paused to get that out the door.  Then I got
distracted with Raptor.

The current checkin I hope will change to use the git repository of
active record constraints directly.  It appears that I have modified
all the migrations to use the new features.  We have
app/models/pg_attribute.rb and others like that.  Those are going to
disappear.  We also have unit tests that are working as well as a
vendor plugin called postgres_constraints which is needs to be
replaced too.

Understand?