Moving towrds better interface
This is a pausing point to branch and see if I can get spry working
with Condor.  Left to do is to finallize hooking up the clip board
stuff.  I'm going to leave that for later.  This is not a usuable
version in the current state but its close.  The clip board list item
could be backed out or just ignored.