This is working well enough to use.
There are a few things that need to be looked in to.

1) I had to not use xhtml because of the embed of ZeroClipboard.  I
think the direction I want to go is towards a unobtrusive technique.
It sounds like serving xhtml is a pain.

2) We path of the json and the html are exactly the same.  Thats going
to make caching a problem.  I don't like the suffix because it any of
the entities have dots in the (like filenames).  I think what I want
to do is create a json namespace.